Fix for use in batch generation.
The name of the output file can now be set using the normal event loop
options so this will work with captControl in batch jobs.  Part of the
change was to move the creation of the output tree into the Initialize
method, and to make sure that events are not saved into the output file.